refactor: 字幕解析结论收拢为跨进程契约类型
